Fixed did_doc crate lints
