+++ did_doc/src/schema/did_doc.rs
@@ -306,14 +306,11 @@ mod tests {
         let service_id = Uri::new("did:example:123456789abcdefghi;service-1").unwrap();
         let service_type = "test-service".to_string();
         let service_endpoint = "https://example.com/service";
-        let service = ServiceBuilder::<()>::new(
-            service_id,
-            service_endpoint.try_into().unwrap(),
-            Default::default(),
-        )
-        .add_service_type(service_type)
-        .unwrap()
-        .build();
+        let service =
+            ServiceBuilder::<()>::new(service_id, service_endpoint.try_into().unwrap(), ())
+                .add_service_type(service_type)
+                .unwrap()
+                .build();
 
         let document = DidDocumentBuilder::new(id.clone())
             .add_also_known_as(also_known_as.clone())
