+++ did_doc/src/schema/service.rs
@@ -226,7 +226,7 @@ mod tests {
 
     #[test]
     fn test_service_serde() {
-        let service_serialized = r##"{
+        let service_serialized = r#"{
           "id": "did:sov:HR6vs6GEZ8rHaVgjg2WodM#did-communication",
           "type": "did-communication",
           "priority": 0,
@@ -238,7 +238,7 @@ mod tests {
             "didcomm/aip2;env=rfc19"
           ],
           "serviceEndpoint": "https://example.com/endpoint"
-        }"##;
+        }"#;
 
         let service: Service<ExtraSov> = serde_json::from_str(service_serialized).unwrap();
         assert_eq!(
